V0.2.1

+++ compiler/ast/ast.c
@@ -8,7 +8,7 @@ Node *node_clone(const Node *const node) {
     clone->type = node->type;
     clone->children_len = node->children_len;
     clone->children = malloc(clone->children_len * sizeof(Node));
-    for (int i = 0; i < clone->children_len; i++) {
+    for (size_t i = 0; i < clone->children_len; i++) {
         clone->children[i] = node_clone(node->children[i]);
     }
     clone->name_len = node->name_len;
@@ -19,7 +19,7 @@ Node *node_clone(const Node *const node) {
 }
 
 
-static Node *new_Node_0(AST nodetype, Token type, char *name /* OWN */, int64_t name_len, int line) {
+static Node *new_Node_0(AST nodetype, Token type, char *name /* OWN */, size_t name_len, size_t line) {
     Node *node = malloc(sizeof(Node));
     node->nodetype = nodetype;
     node->type = type;
@@ -31,7 +31,7 @@ static Node *new_Node_0(AST nodetype, Token type, char *name /* OWN */, int64_t 
     return node;
 }
 
-static Node *new_Node_1(AST nodetype, Token type, Node *child /* OWN */, char *name /* OWN */, int64_t name_len, int line) {
+static Node *new_Node_1(AST nodetype, Token type, Node *child /* OWN */, char *name /* OWN */, int64_t name_len, size_t line) {
     Node *node = malloc(sizeof(Node));
     node->nodetype = nodetype;
     node->type = type;
@@ -44,7 +44,7 @@ static Node *new_Node_1(AST nodetype, Token type, Node *child /* OWN */, char *n
     return node;
 }
 
-static Node *new_Node_2(AST nodetype, Token type, Node *child1 /* OWN */, Node *child2 /* OWN */, char *name /* OWN */, int64_t name_len, int line) {
+static Node *new_Node_2(AST nodetype, Token type, Node *child1 /* OWN */, Node *child2 /* OWN */, char *name /* OWN */, int64_t name_len, size_t line) {
     Node *node = malloc(sizeof(Node));
     node->nodetype = nodetype;
     node->type = type;
@@ -58,7 +58,7 @@ static Node *new_Node_2(AST nodetype, Token type, Node *child1 /* OWN */, Node *
     return node;
 }
 
-static Node *new_Node_3(AST nodetype, Token type, Node *child1 /* OWN */, Node *child2 /* OWN */, Node *child3 /* OWN */, char *name /* OWN */, int64_t name_len, int line) {
+static Node *new_Node_3(AST nodetype, Token type, Node *child1 /* OWN */, Node *child2 /* OWN */, Node *child3 /* OWN */, char *name /* OWN */, int64_t name_len, size_t line) {
     Node *node = malloc(sizeof(Node));
     node->nodetype = nodetype;
     node->type = type;
@@ -74,15 +74,15 @@ static Node *new_Node_3(AST nodetype, Token type, Node *child1 /* OWN */, Node *
 }
 
 
-Node *new_ExprStmt(Node *child, int line) {
+Node *new_ExprStmt(Node *child, size_t line) {
     return new_Node_1(N_EXPRSTMT, T_UNKNOWN, child, NULL, 0, line);
 }
 
-Node *new_Block(Node *body, int line) {
+Node *new_Block(Node *body, size_t line) {
     return new_Node_1(N_BLOCK, T_UNKNOWN, body, NULL, 0, line);
 }
 
-Node *new_Body(int line) {
+Node *new_Body(size_t line) {
     return new_Node_0(N_BODY, T_UNKNOWN, NULL, 0, line);
 }
 
@@ -93,23 +93,23 @@ void body_append(Node *const node, Node *const child) {
 }
 
 
-Node *new_FnDecl(Node *params, Node *body, char *name, int64_t name_len, int line) {
+Node *new_FnDecl(Node *params, Node *body, char *name, int64_t name_len, size_t line) {
     return new_Node_2(N_FNDECL, T_UNKNOWN, params, body, name, name_len, line);
 }
 
-Node *new_Return(Node *expr, int line) {
+Node *new_Return(Node *expr, size_t line) {
     return new_Node_1(N_RET, T_UNKNOWN, expr, NULL, 0, line);
 }
 
-Node *new_Call(Node *params, Node *object, int line) {
+Node *new_Call(Node *params, Node *object, size_t line) {
     return new_Node_2(N_CALL, T_UNKNOWN, params, object, NULL, 0, line);
 }
 
-Node *new_Set(Node *collection, Node *key, Node *value, int line) {
+Node *new_Set(Node *collection, Node *key, Node *value, size_t line) {
     return new_Node_3(N_SET, T_UNKNOWN, collection, key, value, NULL, 0, line);
 }
 
-Node *new_Get(Node *collection, Node *value, int line) {
+Node *new_Get(Node *collection, Node *value, size_t line) {
     return new_Node_2(N_GET, T_UNKNOWN, collection, value, NULL, 0, line);
 }
 
@@ -125,7 +125,7 @@ Node *ListComp_get_collection(const Node *const node) {
     return node->children[2];
 }
 
-Node *new_ListComp(Node *expr, Node *iter, Node *cond, int line) {
+Node *new_ListComp(Node *expr, Node *iter, Node *cond, size_t line) {
     return new_Node_3(N_LISTCOMP, T_UNKNOWN, expr, iter, cond, NULL, 0, line);
 }
 
@@ -133,15 +133,15 @@ Node *TableComp_get_key_value(const Node *const node) {
     return node->children[0];
 }
 
-Node *new_TableComp(Node *expr, Node *iter, Node *cond, int line) {
+Node *new_TableComp(Node *expr, Node *iter, Node *cond, size_t line) {
     return new_Node_3(N_TABLECOMP, T_UNKNOWN, expr, iter, cond, NULL, 0, line);
 }
 
-Node *new_LetIter(Node *var, Node *collection, int line) {
+Node *new_LetIter(Node *var, Node *collection, size_t line) {
     return new_Node_2(N_LETITER, T_UNKNOWN, var, collection, NULL, 0, line);
 }
 
-Node *new_Iter(Node *var, Node *collection, int line) {
+Node *new_Iter(Node *var, Node *collection, size_t line) {
     return new_Node_2(N_ITER, T_UNKNOWN, var, collection, NULL, 0, line);
 }
 
@@ -157,7 +157,7 @@ Node *ForIter_get_body(const Node *const node) {
     return node->children[1];
 }
 
-Node *new_ForIter(Node *iter, Node *body, int line) {
+Node *new_ForIter(Node *iter, Node *body, size_t line) {
     return new_Node_2(N_FORITER, T_UNKNOWN, iter, body, NULL, 0, line);
 }
 
@@ -169,19 +169,19 @@ Node *While_get_body(const Node *const node) {
     return node->children[1];
 }
 
-Node *new_While(Node *cond, Node *body, Node *post, int line) {
+Node *new_While(Node *cond, Node *body, Node *post, size_t line) {
     return new_Node_3(N_WHILE, T_UNKNOWN, cond, body, post, NULL, 0, line);
 }
 
-Node *new_Break(int line) {
+Node *new_Break(size_t line) {
     return new_Node_0(N_BREAK, T_UNKNOWN, NULL, 0, line);
 }
 
-Node *new_Continue(int line) {
+Node *new_Continue(size_t line) {
     return new_Node_0(N_CONT, T_UNKNOWN, NULL, 0, line);
 }
 
-Node *new_If(Node *cond, Node *then_node, Node *else_node, int line) {
+Node *new_If(Node *cond, Node *then_node, Node *else_node, size_t line) {
     return new_Node_3(N_IF, T_UNKNOWN, cond, then_node, else_node, NULL, 0, line);
 }
 
@@ -189,7 +189,7 @@ Node *Print_get_expr(const Node *const node) {
     return node->children[0];
 }
 
-Node *new_Print(Node *expr, int line) {
+Node *new_Print(Node *expr, size_t line) {
     return new_Node_1(N_PRINT, T_UNKNOWN, expr, NULL, 0, line);
 }
 
@@ -197,7 +197,7 @@ Node *Let_get_expr(const Node *const node) {
     return node->children[0];
 }
 
-Node *new_Let(char *name, int64_t name_len, Node *expr, int line) {
+Node *new_Let(char *name, int64_t name_len, Node *expr, size_t line) {
     return new_Node_1(N_LET, T_UNKNOWN, expr, name, name_len, line);
 }
 
@@ -205,47 +205,47 @@ Node *Const_get_expr(Node *node) {
     return node->children[0];
 }
 
-Node *new_Const(char *name, int64_t name_len, Node *expr, int line) {
+Node *new_Const(char *name, int64_t name_len, Node *expr, size_t line) {
     return new_Node_1(N_CONST, T_UNKNOWN, expr, name, name_len, line);
 }
 
-Node *new_TriOp(Token op, Node *left, Node *middle, Node *right, int line) {
+Node *new_TriOp(Token op, Node *left, Node *middle, Node *right, size_t line) {
     return new_Node_3(N_TRIOP, op, left, middle, right, NULL, 0, line);
 }
 
-Node *new_BinOp(Token op, Node *left, Node *right, int line) {
+Node *new_BinOp(Token op, Node *left, Node *right, size_t line) {
     return new_Node_2(N_BINOP, op, left, right, NULL, 0, line);
 }
 
-Node *new_UnOp(Token op, Node *child, int line) {
+Node *new_UnOp(Token op, Node *child, size_t line) {
     return new_Node_1(N_UNOP, op, child, NULL, 0, line);
 }
 
-Node *new_Assign(char *name, int64_t name_len, Node *child, int line) {
+Node *new_Assign(char *name, int64_t name_len, Node *child, size_t line) {
     return new_Node_1(N_ASSIGN, T_UNKNOWN, child, name, name_len, line);
 }
 
-Node *new_Var(char *name, int64_t name_len, int line) {
+Node *new_Var(char *name, int64_t name_len, size_t line) {
     return new_Node_0(N_VAR, T_UNKNOWN, name, name_len, line);
 }
 
-Node *new_Undef(int line) {
+Node *new_Undef(size_t line) {
     return new_Node_0(N_UNDEF, T_UNKNOWN, NULL, 0, line);
 }
 
-Node *new_Float(char *value, int len, int line) {
+Node *new_Float(char *value, size_t len, size_t line) {
     return new_Node_0(N_FLOAT64, T_UNKNOWN, value, len, line);
 }
 
-Node *new_Integer(char *value, int len, int line) {
+Node *new_Integer(char *value, size_t len, size_t line) {
     return new_Node_0(N_INT64, T_UNKNOWN, value, len, line);
 }
 
-Node *new_Boolean(char *value, int len, int line) {
+Node *new_Boolean(char *value, size_t len, size_t line) {
     return new_Node_0(N_BOOL, T_UNKNOWN, value, len, line);
 }
 
-Node *new_String(char *value, int len, int line) {
+Node *new_String(char *value, size_t len, size_t line) {
     return new_Node_0(N_STR, T_UNKNOWN, value, len, line);
 }
 
@@ -253,7 +253,7 @@ Node *List_get_values(const Node *const node) {
     return node->children[0];
 }
 
-Node *new_List(Node *values, int line) {
+Node *new_List(Node *values, size_t line) {
     return new_Node_1(N_LIST, T_UNKNOWN, values, NULL, 0, line);
 }
 
@@ -261,7 +261,7 @@ Node *Table_get_values(const Node *const node) {
     return node->children[0];
 }
 
-Node *new_Table(Node *keys, int line) {
+Node *new_Table(Node *keys, size_t line) {
     return new_Node_1(N_TABLE, T_UNKNOWN, keys, NULL, 0, line);
 }
